refactor(agent): move third-party agent wrappers into agent/external/
Separate the vendor-SDK localization agents (ClaudeLocAgent, CodexLocAgent)
from CodeMiner's own agent stack by relocating them under a dedicated
agent/external/ package. Fixes relative log_utils imports and the example
import paths to match.
