+++ codeminer/agent/external/claude_agent.py
@@ -21,7 +21,7 @@ from typing import Any, Dict, List, Optional
 
 from claude_agent_sdk import ClaudeAgentOptions, query
 
-from ..log_utils import get_logger
+from ...log_utils import get_logger
 
 # Canonical chunker-format rules for the `name` field. Shared with
 # codex_agent so both system prompts derive from a single source of truth.
